django/requirements.txt
Django>=3.1
psycopg2
uwsgi
requests
Pillow
pandas
plotly
django-ipware
django-cors-headers